Logging

+++ src/main/java/com/elitvinchuck/recipes/controllers/RecipeController.java
@@ -1,13 +1,15 @@
 package com.elitvinchuck.recipes.controllers;
 
+import com.elitvinchuck.recipes.constants.RecipeConstants;
 import com.elitvinchuck.recipes.exceptions.RecipeNotFoundException;
 import com.elitvinchuck.recipes.models.Recipe;
 import com.elitvinchuck.recipes.repositories.RecipeRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.util.Map;
 import java.util.Optional;
@@ -16,7 +18,9 @@ import java.util.Optional;
 @RequestMapping("recipe")
 public class RecipeController {
 
-    private RecipeRepository repository;
+    private static final Logger LOGGER = LoggerFactory.getLogger(RecipeController.class);
+
+    private final RecipeRepository repository;
 
     @Autowired
     public RecipeController(RecipeRepository repository) {
@@ -27,9 +31,11 @@ public class RecipeController {
     public Recipe getRecipe(@PathVariable long id) {
         Optional<Recipe> recipeOptional = repository.findById(id);
         if (recipeOptional.isPresent()) {
+            LOGGER.info(RecipeConstants.GET_RECIPE_SUCCESS_STRING + recipeOptional.get());
             return recipeOptional.get();
         }
-        throw new RecipeNotFoundException("Recipe not found for id=" + id);
+        LOGGER.error("GET: " + RecipeConstants.RECIPE_NOT_FOUND_FOR_ID_STRING + id);
+        throw new RecipeNotFoundException(RecipeConstants.RECIPE_NOT_FOUND_FOR_ID_STRING + id);
     }
 
     @PostMapping("/new")
@@ -37,15 +43,21 @@ public class RecipeController {
         long id = repository
                 .save(recipe)
                 .getId();
+        LOGGER.info(RecipeConstants.POST_RECIPE_SUCCESS_STRING + recipe);
         return Map.of("id", id);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> delete(@PathVariable long id) {
-        repository.deleteById(id);
-        return ResponseEntity
-                .noContent()
-                .build();
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            LOGGER.info(RecipeConstants.DELETE_RECIPE_SUCCESS_STRING + id);
+            return ResponseEntity
+                    .noContent()
+                    .build();
+        }
+        LOGGER.error("DELETE: " + RecipeConstants.RECIPE_NOT_FOUND_FOR_ID_STRING + id);
+        throw new RecipeNotFoundException(RecipeConstants.RECIPE_NOT_FOUND_FOR_ID_STRING + id);
     }
 
 }
